backend/src/requirements.txt
aws-lambda-powertools
fastjsonschema
aws_xray_sdk
bcrypt
PyJWT
requests
urllib3==1.26.6
python-dotenv
alembic
sqlalchemy
psycopg2-binary
pandas
google-auth